pyproject.toml
[project]
name = "weather-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.6",
    "polars>=1.17.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.36",
]
